fix bug #273: terminal download model argument type error (#274)
**Description:**

Please provide a brief description of the changes made in this pull
request.

**Related Issue:**

If this pull request is related to any existing issue, please mention
the issue number here.

**Changes Made:**

Please list down the specific changes made in this pull request.

**Testing Done:**

Please describe the testing that has been done to ensure the changes
made in this pull request are functioning as expected.

**Screenshots:**

If applicable, please provide screenshots or GIFs or videos to visually
demonstrate the changes made.

**Checklist:**

- [x] I have tested the changes locally.
- [x] I have self-reviewed the code changes.
- [x] I have updated the documentation, if necessary.

Co-authored-by: ys <zhaianhua@chinauniwise.com>