Switch llama-cpp backend to use the official llama-cpp vulkan backend release instead of llama-cpp-python (#286)
**Description:**

This PR switches from using a self-built llama-cpp-python to the
official llama-cpp vulkan backend release. This should increase
performance during inference and improve the maintainability of AI
Playground.

The PR also adds the ipex-llm ollama backend as a proof-of-concept
behind a feature flag.

**Changes Made:**

* switch llama-cpp from self-built llama-cpp-python to official
llama-cpp server with vulkan backend
* add deepseek gguf model
* add ollama PoC (behind feature flag)

**Testing Done:**

Tested locally on B580

**Screenshots:**


![image](https://github.com/user-attachments/assets/0ef8cebe-09a9-437d-81ae-44acb0bded1e)

**Checklist:**

- [x] I have tested the changes locally.
- [x] I have self-reviewed the code changes.

---------

Signed-off-by: Markus Schuettler
Co-authored-by: Li-AI-ABN <qiacheng.li@intel.com>